For #9815: "Open in Firefox Preview" from Custom Tab should open tab, not home page.
Since last commit (a477f14) didn't call `match()`, TabIntentProcessor won't have the chance to intercept the intent from Custom Tab "Open in Firefox Preview" menu item.

Before we fix the combination of `match()` and `process()` in A-C, we need to call it this way here: if matched, then process it.